+++ spec/models/user_spec.rb
@@ -1,5 +1,0 @@
-require "rails_helper"
-
-describe User do
-  it { is_expected.to validate_presence_of :full_name }
-end